chore: Phase 1b - move configure-v3-client/insomnia-fetch behind insomnia-api
Relocates configureV3ClientDefaults, insomniaFetch/proxyAwareFetch, and
the four small constants they depend on (getApiBaseURL, getClientString,
INSOMNIA_FETCH_TIME_OUT, PLAYWRIGHT_TEST) from packages/insomnia/src into
insomnia-api, which both desktop and the CLI already depend on. Desktop's
own common/constants.ts, common/insomnia-fetch.ts and
common/configure-v3-client.ts become thin re-export shims (same pattern
already used for generateId in misc.ts), so every existing desktop import
path keeps working unchanged.

insomnia-inso's cli.ts now imports configureV3ClientDefaults/insomniaFetch
from insomnia-api directly instead of reaching into packages/insomnia/src.

getClientString's version string now reads insomnia-api's own package.json
instead of desktop's - safe because every workspace package in this
monorepo is released with the same synchronized version number, so the
resulting string is identical today and stays identical going forward.

send-request.node.ts's relocation is deliberately deferred (curl/scripting/
vault coupling makes it a separate, higher-risk piece of work) - the CLI
still reaches into packages/insomnia/src/network/send-request.node for now.

Verified via insomnia-api/insomnia-inso/insomnia unit tests (139 test files,
2175 tests passing in packages/insomnia alone) and the full inso dev-bundle
e2e suite, including tests that exercise the real network path
(insomnia.sendRequest, run collection HTTP requests) - all pass unchanged.
