+++ packages/insomnia-api/src/index.ts
@@ -11,3 +11,6 @@ export * from './mock';
 export * from './vcs';
 
 export { configureFetch, type FetchConfig, ResponseFailError, isApiError } from './fetch';
+export { getApiBaseURL, getClientString, INSOMNIA_FETCH_TIME_OUT, PLAYWRIGHT_TEST } from './client-defaults';
+export { insomniaFetch, proxyAwareFetch, setFetchImplementation } from './insomnia-fetch';
+export { configureV3ClientDefaults } from './configure-v3-client-defaults';
